Fixing diffSec in TimeStamp

+++ timing/TimeStamp.cpp
@@ -30,8 +30,8 @@ namespace Utils {
 double diffSec(const Utils::Timing::TimeStamp & src,
                const Utils::Timing::TimeStamp & dst)
 {
-  int elapsedTicks = (dst - src).count();
-  return (double)(elapsedTicks) * steady_clock::period::num / steady_clock::period::den;
+  double elapsedTicks = (dst - src).count();
+  return elapsedTicks * steady_clock::period::num / steady_clock::period::den;
 }
 
 double diffMs(const Utils::Timing::TimeStamp & src,
